+++ netbox_ai_navigator/model_providers/__init__.py
@@ -1,11 +1,9 @@
 from .base import ModelProvider, ModelResponse, ModelToolCall
-from .mygpt_api import MyGPTApiProvider
 from .openai_compatible import OpenAICompatibleProvider
 
 __all__ = (
     "ModelProvider",
     "ModelResponse",
     "ModelToolCall",
-    "MyGPTApiProvider",
     "OpenAICompatibleProvider",
 )